# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Market Backend API
  version: 0.1.0
  description: 'Unified HTTP API contract for Market Backend. Operations marked with `x-status: implemented` are
    exposed by the current FastAPI application; operations marked with `x-status: planned` are contract-first proposals
    and are not available yet.'
paths:
  /api/v1/admin/auth/token:
    post:
      tags:
      - admin
      summary: Login
      operationId: login_api_v1_admin_auth_token_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
...
  /api/v1/admin/tags:
    get:
      tags:
      - admin
      summary: List Tags
      operationId: list_tags_api_v1_admin_tags_get
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagResponse'
                title: Response List Tags Api V1 Admin Tags Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
    post:
      tags:
      - admin
      summary: Create Tag
      operationId: create_tag_api_v1_admin_tags_post
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
  /api/v1/admin/tags/{tag_id}:
    patch:
      tags:
      - admin
      summary: Update Tag
      operationId: update_tag_api_v1_admin_tags__tag_id__patch
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Tag Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
    delete:
      tags:
      - admin
      summary: Delete Tag
      operationId: delete_tag_api_v1_admin_tags__tag_id__delete
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Tag Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
  /api/v1/admin/orders:
    get:
      tags:
      - admin
      summary: List Orders
      operationId: list_orders_api_v1_admin_orders_get
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrmOrderResponse'
                title: Response List Orders Api V1 Admin Orders Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
  /api/v1/admin/orders/{order_id}:
    patch:
      tags:
      - admin
      summary: Update Order Status
      operationId: update_order_status_api_v1_admin_orders__order_id__patch
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmOrderUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrmOrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
...
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Register
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
  /api/v1/auth/token:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_token_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
...
  /api/v1/orders:
    post:
      tags:
      - orders
      summary: Create Order
      operationId: create_order_api_v1_orders_post
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
    get:
      tags:
      - orders
      summary: List Orders
      operationId: list_orders_api_v1_orders_get
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
                title: Response List Orders Api V1 Orders Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
  /api/v1/orders/{order_id}:
    get:
      tags:
      - orders
      summary: Get Order
      operationId: get_order_api_v1_orders__order_id__get
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
...
components:
  schemas:
    AdminLoginRequest:
      properties:
        email:
          type: string
          maxLength: 320
          minLength: 1
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
      additionalProperties: false
      type: object
      required:
      - email
      - password
      title: AdminLoginRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        email:
          type: string
          maxLength: 320
          minLength: 1
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
      additionalProperties: false
      type: object
      required:
      - email
      - password
      title: LoginRequest
...
    RegisterRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
        phone:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Phone
      additionalProperties: false
      type: object
      required:
      - email
      - password
      title: RegisterRequest
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
      additionalProperties: false
      type: object
      required:
      - access_token
      - token_type
      title: TokenResponse
...
    UserResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        email:
          type: string
          format: email
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        telegram_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Telegram Id
        telegram_username:
          anyOf:
          - type: string
          - type: 'null'
          title: Telegram Username
        is_active:
          type: boolean
          title: Is Active
        is_superuser:
          type: boolean
          title: Is Superuser
        role:
          type: string
          enum:
          - user
          - moder
          - admin
          - developer
          - owner
          title: Role
        is_email_verified:
          type: boolean
          title: Is Email Verified
        is_phone_verified:
          type: boolean
          title: Is Phone Verified
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      additionalProperties: false
      type: object
      required:
      - id
      - username
      - email
      - phone
      - telegram_id
      - telegram_username
      - is_active
      - is_superuser
      - role
      - is_email_verified
      - is_phone_verified
      - created_at
      - updated_at
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    CreateOrderItemRequest:
      properties:
        product_id:
          type: string
          format: uuid
          title: Product Id
        quantity:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Quantity
      additionalProperties: false
      type: object
      required:
      - product_id
      - quantity
      title: CreateOrderItemRequest
    CreateOrderRequest:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CreateOrderItemRequest'
          type: array
          maxItems: 100
          minItems: 1
          title: Items
        customer:
          $ref: '#/components/schemas/CustomerRequest'
        comment:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Comment
      additionalProperties: false
      type: object
      required:
      - items
      - customer
      title: CreateOrderRequest
...
    CustomerRequest:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        email:
          type: string
          maxLength: 320
          format: email
          title: Email
        phone:
          type: string
          maxLength: 32
          minLength: 1
          title: Phone
      additionalProperties: false
      type: object
      required:
      - name
      - email
      - phone
      title: CustomerRequest
    CustomerResponse:
      properties:
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        phone:
          type: string
          title: Phone
      additionalProperties: false
      type: object
      required:
      - name
      - email
      - phone
      title: CustomerResponse
    OrderItemResponse:
      properties:
        product_id:
          type: string
          format: uuid
          title: Product Id
        title:
          type: string
          title: Title
        quantity:
          type: integer
          title: Quantity
        unit_price:
          type: string
          title: Unit Price
        total:
          type: string
          title: Total
      additionalProperties: false
      type: object
      required:
      - product_id
      - title
      - quantity
      - unit_price
      - total
      title: OrderItemResponse
    OrderResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        status:
          type: string
          title: Status
        items:
          items:
            $ref: '#/components/schemas/OrderItemResponse'
          type: array
          title: Items
        total:
          type: string
          title: Total
        currency:
          type: string
          title: Currency
        customer:
          $ref: '#/components/schemas/CustomerResponse'
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      additionalProperties: false
      type: object
      required:
      - id
      - status
      - items
      - total
      - currency
      - customer
      - comment
      - created_at
      - updated_at
      title: OrderResponse
...
    CrmOrderResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        customer_name:
          type: string
          title: Customer Name
        customer_email:
          type: string
          format: email
          title: Customer Email
        total:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total
        status:
          $ref: '#/components/schemas/CrmStatus'
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - customer_name
      - customer_email
      - total
      - status
      - created_at
      title: CrmOrderResponse
    CrmOrderUpdateRequest:
      properties:
        status:
          $ref: '#/components/schemas/CrmStatus'
      additionalProperties: false
      type: object
      required:
      - status
      title: CrmOrderUpdateRequest
    CrmStatus:
      type: string
      enum:
      - payment_verification
      - assembling
      - ready_to_ship
      - in_transit
      - awaiting_pickup
      - received
      - closed
    TagCreateRequest:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        slug:
          type: string
          maxLength: 100
          minLength: 1
          title: Slug
        description:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Description
      additionalProperties: false
      type: object
      required:
      - name
      - slug
      title: TagCreateRequest
    TagResponse:
      properties:
        uuid:
          type: string
          format: uuid
          title: Uuid
        name:
          type: string
          title: Name
        slug:
          type: string
          title: Slug
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      additionalProperties: false
      type: object
      required:
      - uuid
      - name
      - slug
      - description
      - created_at
      - updated_at
      title: TagResponse
...
    TagUpdateRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Name
        slug:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Slug
        description:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Description
      additionalProperties: false
      type: object
      title: TagUpdateRequest
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
servers:
- url: /
  description: Current origin
...